requirements_embedding.txt
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.21.0
scikit-learn>=1.0.0 